Extract parse_result into new module
